Allow 32-bit registers in `classical.expr` system
The current limit in `Type` limits `Uint` sizes to 16-bit (65,535).  We
want to do scaling tests storing one million or more bits, so we need to
widen the type.  QPY already uses a 32-bit width in this location, and
the 16-bit internal limit was just a memory optimisation.

+++ crates/qpy/src/value.rs
@@ -872,7 +872,7 @@ fn pack_expression_type(exp_type: &Type, version: u32) -> Result<ExpressionType,
                 })
             }
         }
-        Type::Uint(width) => Ok(ExpressionType::Uint(*width as u32)),
+        Type::Uint(width) => Ok(ExpressionType::Uint(*width)),
     }
 }
